+++ src/instance/state.rs
@@ -8,8 +8,6 @@ use ::tarantool::tlua;
         /// Instance has gracefully shut down or has not been started yet.
         #[default]
         Offline = "Offline",
-        /// Instance has configured replication.
-        Replicated = "Replicated",
         /// Instance is active and is handling requests.
         Online = "Online",
         /// Instance has permanently removed from cluster.